パスワード入力部分のViewを別ファイルに抽出
