+++ SwiftUIStudy/View/SecureTextField/SecureTextField.swift
@@ -29,35 +29,6 @@ struct SecureTextField: View {
     }
 }
 
-struct PasswordView: View {
-
-    @State private var password: String = ""
-    @State private var isSecured: Bool = true
-
-    var body: some View {
-        HStack {
-            if self.isSecured {
-                SecureField("Password", text: $password)
-                    .textFieldStyle(RoundedBorderTextFieldStyle())
-                    .padding(.leading, 20.0)
-            } else {
-                TextField("Password", text: $password)
-                    .textFieldStyle(RoundedBorderTextFieldStyle())
-                    .padding(.leading, 20.0)
-            }
-            Button(action: {
-                self.isSecured.toggle()
-            }) {
-                self.isSecured ?
-                    Image(systemName: "eye.slash.fill") :
-                    Image(systemName: "eye.fill")
-            }
-            .foregroundColor(.gray)
-            .padding(.trailing, 20.0)
-        }
-    }
-}
-
 struct LoginButton: View {
     var body: some View {
         Button(action: {
